requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
sqlalchemy>=2.0.0
pydantic>=2.4.2
pydantic-settings>=2.0.3
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1
python-multipart>=0.0.6
email-validator>=2.0.0
aiokafka>=0.8.1
loguru>=0.7.0
prometheus-fastapi-instrumentator>=6.1.0
slowapi>=0.1.8
pytest>=7.4.2
httpx>=0.25.0
psycopg2-binary>=2.9.9
